List EVC tree once only when querying with name
Why:

The children should not be listed multiple times. If they appear in the
EVC tree of some parent experiment then they should not be listed as a
root as well.

How:

Problem was that all experiments matching the name would be fetched and
printed with their tree. When name is specified, version queried should
be 1 by default unless specified by the user.