Fix TPE sampling in narrow spaces
Why:

The GMM for real values could not sample efficiently in narrow spaces.
It would often lead to RuntimeError because the number of attempts
allowed would be exhausted. We could increase the default number of
attempts allowed but that would increase the computational cost for any
space, even those easy to sample.

How:

Use numpy array to avoid playing with a list. It is more efficient.
Also, increase the number of attempts as needed until it reaches a max
value of attempts. This way easy samples do not take more time while
difficult ones are allowed more.