Clear error message for dup branching error
Why:

When trying to branch from an experiment that already has a child with
the same name, Oríon will crash with a RaceCondition error. The problem
is that this issue and a real race-condition are indiscernible as they
lead to the same state. The only thing we can do is clarify the error
message and warn that this error can also be caused by branching from a
parent experiment that already has a child with the same name.

Note:

This issue only arises if the user specifies the version of the parent
experiment, otherwise the EVC will use the child or branch from the
child without any issue.