Fix shape argument in categorical dims
Why:

The shape of categorical dimensions was not included in
`get_prior_string`, causing the lost of the shape during branching.

How:

Add the shape to `get_prior_string` of Categorical dimensions and add
tests to catch this issue.

Also add tests for Conflict and Resolutions of priors with different
shape.

The adaptor for a change of prior does not raise an issue anymore when
shapes are different and rather logs a warning. The trials are all
ignored is this case.