Use user_script_config from parser in EVC
Why:

The option `user_script_config` is part of the worker configuration, not
the EVC. As such, this option is not part of `branching` group of option
and does not find its way to the Conflict objects of the EVC. The value
of user_script_config is already available anyway inside the
experiment configuration, so there is no need to pass it. Furthermore, it may
differ from past experiments to the new one, and both experiments should be handled
based on their respective `user_script_config`, not the new one. For
these reasons, it is better to use the information available in the
experiment configurations.