Support linearize log integer properly
Why:

A log integer dimension would be casted to real, linearized, then casted
back to integer. This reduces dramatically the number of possible values
that can be sampled as many exp(int(log(x))) will result in the same
integer for many different values of x. An algorithm that needs
linearization should be able to handle real space or otherwise state a
requirement for integers. This should be handled separately and
quantization of linearized log integer should not be applied by default.

Note:

Due to the use of floor instead of rounding in Quantize, the values of
int(exp(log(x))) would still clash for close values of x. Using rounding
instead solves the issue. Rounding may be problematic however for
algorithms that require integer type, as the rounding may cast real
integers to values that are out-of-bound. For now there are no
foreseeable algorithms that may require integer type so I avoid fixing
the issue and leave it for later if the need even arises (which I highly
doubt).