Fix experiment tree tests
Why:

There was a bug in the tests. The functions to generate trials would
generate more than requested because of the new behavior of producer
attempting to produce all trials at once, once the value of `max_trials`
was conflicting with the number of trials requested to the trial
generating function for tests (`orion.testing.evc.generate_trials`).

Fortunately the bug in the tests did not seem to miss any bugs in the
code they were testing.

How:

Adjust the expected numbers based on the corrected behiavor. The numbers
make indeed more sense now.