Add new Factory classes
Why:

The meta-class factory was confusing and sometimes hard to debug. The
factory should be a separate object that the constructors of the
classes. When calling a class, we expect to receive an instance of the
class, not of a child class. The factory also makes it simpler to add
new features such as fetching class objects based on strings on grouping
singleton instances per factory.

How:

Each class requiring the factory will be using ``Factory(base_class)``
as a factory to instantiate child classes. The explicit method
``factory.create()`` will be used instead of the confusing
`MyFactory()`.

Co-authored-by: François Corneau-Tremblay <corneau90@gmail.com>