Adjust Transformers space to using Trials
Why:

The algorithms will be working with trial objects instead of tuple of
values. This means the space needs to sample trials, and space
transformations should be applied on trials as well instead of tuples of
values.

How:

For simplicity, only interface of the space classes (TransformedSpace
and ReshapedSpace) will be working with trials. The transformations per
dimension will be applied using tuple of values so that, in particular,
reshaping operations remain straightforward.

To facilitate debugging, transformed trials are wrapped so that the
original trial can still be accessible. This will prove handy in
algorithms if we need access to original trial objects, and also because
the ID of the transformed trial should always be based on the original
parameters (otherwise the ID gets incoherent with the database).