Handle reshaped dimension with diff. name ordering
Why:

If the names of the dimensions have the same prefix, but some have a
shape, a transformed space with have a different ordering of dimension.
For example, the following dimensions will have their named sorted
differently:

  dim (shape 2), dim1 (no shape) -->  dim1, dim[0], dim[1]

This is causing an issue when we try to restore the shape of the
transformed dimension, with the names being swapped.

How:

When restoring shape, keep track of the original keys and their order,
and reassign the restored dimensions to the correct index (correct dim
name).