Support reshaped transform on dim + default_value
Reshaped transform on dim:

Sometimes we need to apply a transformation on a dimension only. So far
the transformation of View could only be applied on a full point,
because the ReshapedDimension wrapping the view was expected to fetch
the corresponding dimension at self.index. This commit pushes up the
fetch at self.index at the level of the ReshapedSpace instead, the only
place where we are expected to work on a full point, not only on a
dimension.

default value:

Transformed dimensions should support default value like original
dimensions.