Improve resiliency of trial reservation
Why:

We had two levels of patience when reserving a trial. There was the
customizable `max_idle_time` used in producer.produce() to limit the
time spend trying to generate new trials, and there was `_max_depth` in
`reserve_trial` limiting the number of times a reservation would be
attempted and producer.produce would be called. This lead to misleading
error messages. For instance, with many workers it happened that a
worker would always be unable to reserve a trial because each time it
executed producer.produce() all other workers would reserve the trials
before the current worker had time to reserve one. In such scenario the
error message would be that the algorithm was unable to sample new point
and is waiting for trials to complete. It is not true. We should state
the number of trials that were generated during these reservation
attempts and recommend increasing the pool-size and timeout.

How:

Producer.produce only attempts producing `pool-size` once (calling
algo.suggest only once) and returns the number of successfully
produced trials. The whole patience is moved to `reserve_trial` where
it attempts reserving and producing until it reaches the timeout, in
which case a helpful error message is raised.