Infer trial working dir based on exp.working_dir
Why:

The trial working dir should be unique to the trial and depend on the
experiment's working dir. We can use the id of the trial (or variants
ignoring fidelity, experiment id or lies) to define a unique working
dir.

How:

Instead of setting the working dir directly, we set the experiment's
working dir.