Fix ASHA for identical objectives
Why:

The sorting of trials during promotion was not only looking at
objectives, but also at trials. Problem is, trials are not comparable.
This bug was not detected because during tests trials always have a
different objective, hence the comparisons would always stop at the
objective when comparing the tuples (objective, trial).

How:

Add tests with identical objectives, and use objective only for sorting.