Deprecate parallel strategy configuration
The parallel strategy will now be part of algorithm configuration. There
is nothing to configure in the producer so it is removed altogether from
experiment configuration.

Any configuration passed for parallel strategy will trigger a warning
and be ignored. It would be too complicated to try to map this to
algorithm arguments for parallel strategies.

Deprecation is planned for v0.4.