Add BatchWrite to batch DB operations during lock
Why:

When we lock the algorithm, any write of new trial should not persist in
the database if the execution crashes during the lock. In such case the
algorithm's state will be reverted and keeping the new trials in the DB
would lead to an inconsistent state between trials generated by algo and
algorithm state.