Add acquire_algorithm_lock to storage
Why:

The algorithm state must be locked by one runner at a time during
experiment execution. Instead of replicating the algorithms, we will
share the state and whenever a runner needs to sample new trials, it can
lock and load the current state of the algorithm, sample new trials and
then save back the new state of the algorithm before releasing the lock.
This way we have something equivalent to having a single algorithm
running in a master process, but it can be executed by any runner,
making the whole system more resilient if we have multiple runners.

How:

The storage has a new method acquire_algorithm_lock, a context_manager,
which locks the state in the storage during during `__enter__` and
release the lock during `__exit__`.