Adapt Producer to shared state algorithm
The producer is not much more simpler. It simply serve as a bridge
between the algorithm and the database. When calling observe or produce,
it acquires the lock on the algorithm, update the algorithm (observe or
suggest new points) and then let the experiment save back the new state
of the algorithm and release the lock.