Avoid raising LazyWorkerIdle if observe is slow
If observe is slow (which is possible if acquiring the algorithms lock
takes time), then all trials may complete during the gather and the
runner will start counting the idle time of the workers.
It should only start counting the idle time if nothing happened during a
loop, that is, if there were no trials to scatter and nothing gathered.