Reduce DB I/O during optimization
Runner was calling the DB at every loop, through `experiment.is_done`.
This was causing huge I/O, up to 2MB/s for a single Runner. These calls
were not necessary since we could look at `self.free_worker` first to
determine if it was necessary to sample new trials. This PR reduces the
I/O level about 78 times less for working history of ~500 trials and
trial executions of ~30mins. There is still a lot of room for
improvement for the efficiency of `is_done` calls and algorithm locking
mechanism.

A new test is added to ensure that further changes do not increase the
I/O levels significantly above current nominal levels (~17KB/s in and
~59KB/s out for the test case).