Remove Benchmark {algorithm: algo, deterministic:}
Why:

The interface is bloated and difficult to memorise. One of the reason we
decided to go this way is that we previously needed to instantiate an
algo to access its attributes. We can now use algo_factory.get_class()
to have the class object only and access the class attributes. Thanks to
this we can now simply set deterministic as a class attribute for the
algorithms and it simplifies the interface of the benchmark.

Algorithms have deterministic set to False by default since they
generally have sources of variations.