Improve prior build error message
If there is an error during the attempt of building a builtin prior
(methods implemented in SpaceBuilder), the builder assumes the
expression is not a builtin prior and then tries to create a scipy.stat
prior which also fails. The error message is then that the name of the
expression is not a valid prior. We should instead explicitly verify if
the name of the prior is one that is builtin and then try to build it
without silencing error messages. This way we get the correct error
message for the prior.