Fix bug with algo param_names (#947)
Fix bug with algo param_names (see desc.)

In python 3.7, the Generic class has a metaclass (GenericMeta or
something similar), and inspect.signature(type(self)) would give
the signature of the metaclass's `__call__` method, instead of
the signature of the __init__ of the class.

Also fixed a bug whereby it would include "self" and "args" from
*args into the param names.
This wasn't causing issues because we check if `self` has an
attribute with the given name before including it in the
configuration.

Signed-off-by: Fabrice Normandin <fabrice.normandin@gmail.com>