Do not hash experiment for trial.id
It is problematic to include experiment in trial hash since the
experiment it belongs to is not known in advance in the algorithm
object. In any case, we can make a unique index combining trial.hash +
experiment.id so it is not necessary to hash the experiment.id.

The trial has now trial.id, the hash of params+parent+fidelity and
trial.id_override, the id corresponding to _id in legacy databases.
trial.legacy_id is added to keep a backward compatible hash with
previous versions of Oríon where trial.id included the experiment id in
the hash.