Fix ephemeraldb mixed index types
Why:

When upgrading the DB to support index of (trial.id, trial.experiment),
the DB will end up with a mixed format for the trial ids. The previous
ids (_id) where the hash of the trials. It will now rely on the DB to
generate the next _id. Thus, it will have string index mixed with
integer values. It is best to keep ids as they were to better support
backward compatibility (and Oríon version down-grade), so we should
support mixed indexes. The bug was confusing anyway.

How:

Only consider integer indexes when inferring next id to use.