Add prepare_trial arg to Runner and exp.workon
Why:

When the trial has a parent, it means it should start from the same
working_dir state. We need to fetch the parent trial and copy its working
dir to the current trial's working dir.

How:

The runner now has a new argument, a callable that will be executed
before a trial is submitted to the executor. By default this callable
will take care of copying parent trial working_dir to child trial
working dir, or simply create the working dir if the trial has no
parent.

If the callable fails, the exception is caught and delayed to be caught
again during the executor's execution of the trial. This makes it
possible to handle these trials the same way as if they crashed during
trial execution rather than during trial preparation.