Handle properly parent id in AlgoWrapper
Why:

The parent id is in transformed space. When converting a trial from
transformed space to original space, the id of the parent must be
converted too. If the parent has a parent too, then this grand-parent
must be converted too, and so on.

How:

Recursively get to the root trial, and then compute the trial ids down
to last parent trial.

Tests are added to verify this.