Fix bug in test-suite
Why:

The phase method was not defined as a `classmethod`. Because of this,
the current_phase was set on the instance of the class but not on the
class itself. This was problematic inside `create_algo` which is a
`classmethod` and therefore `_current_phase` was not the current one, as
it was set on the instance but not the class.