Use random objectives instead of incremental
Why:

The reproducibility bug in PB2 was not detected because it was
systematically selecting the limit close to first trial, because it was
the closest to best objective. By sampling the objective randomly it is
more likely to detect different behaviors and bugs. The seeds are fixed
thought, to avoid flaky tests, but we can parametrize them.

Note:

No tests relied on the incremental objective, but it did uncover bugs in
other algorithms. One confusing thing is that ASHA and TPE tests
test_suggest_n now sample less than `n`. It is normal for ASHA, is it
now reaches last rung that is smaller than `n`, but it is not clear for
TPE. TPE at phase `bo` should indeed sample only 1 trial at a time, but
the wrapper should be able to request trials many times and accumulate
`n` as requested. To be investigated...