Relax RNG seed test
It should not assume that algorithms are able to sample 2 trials at the
same time. Therefore we won't test that it returns different trials one
after the other. Only that seeding gives the same trial. The convergence
test will figure it out if the algorithm always return the same trial.