Make sure to completely reload experiments pages when selected experiment changes.
Previously, a same component was used and reloaded each time an experiment is selected.
So, if we select an experiment, then another one just after without waiting for the former to load, thus the former may replace the later in the component when loading request is completed.

To prevent this, we make sure a fully new component is recreated each time an experiment is selected.